requirements.txt

langchain_community
langchain-ollama
streamlit>=1.46.0
pandas>=2.3.0
numpy>=2.0.0
plotly>=6.1.2
requests>=2.32.4
google-api-python-client>=2.172.0
google-auth>=2.40.3
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
